gui/Inktool.py : on-canvas editing node pressure feature added.
Currently,first of all,place pointer on the node you want to edit
and hold pressing shift key,then drag node.
then the node position remained,only pressure is changed.

when drag up or left,it decrease pressure of the node.
when drag down or right,it increase pressure of the node.

also,in some environment,you can change a node pressure with mouse wheel.