ProjectSave: bug fixed & codes reorganized.
ProjectSave has a bug,when opened work is not
project and then 'save as project',but Projectsave
cannot change name of current work.
under this situation, if 'save'(overwrite) menu executed,
it overwrite as old filename,not as project.
this bug fixed.

and codes reorganized, especially project-save related method
options handling changed.