Stabilizer: rare case bug found.
Rare case bug found, but I'm not sure
how to reproduce it.
By the way, in such case, stabilizer
enumerates resampled cursor position as
(None, None), so tdw.display_to_model()
failed.

In this case, pressure value was 0,
and I think it might be the line

    yield (self._cx, self._cy, 0.0)

    or

    yield (self._rx, self._ry, 0.0)

returned (None, None, 0.0)

so, I rewrite some lines to initialize these valus with 0.0,
not None.