ExInkingMode & BezierMode: Now added 'node insert' feature.
Now ExInkingMode and BezierMode can insert a node
into currently editing stroke.

Using new Gtk.Action 'InsertNodePhase' (from keyboard binding),
we can insert a node into the current stroke, with clicking
on it.

That phase is one-shot. Regardless whether success to insert node or not,
the phase of mode return to adjusting phase.