gui/keyboard.py: fixed modifier only event
When there is a button mapping such as 'CTRL + BUTTON1'
(as default, it should be color-picker)
and there is a shortcut such as 'CTRL + S',
we likely to hold CTRL key prior to S key.
Then after Gtk.Action for CTRL+S fired, CTRL keyup event
is not invoked, so color-picker remained still activated.

This bug fixed.