gap-closing flood-fill: Initial commit.
Implemented psuedo Gap-closing flood-fill feature.
And, FloodFillMode(FloodFillOptionsWidget) has new 2 options,
which is 'Dilation Size" and "Gap Closing Radius".
'Dilation Size' is to assign how many pixels flood-filled area
to be dilated.
'Gap Closing Radius' is, to assign how many pixels to close
contour gap of the area to be flood-filled.This is radius, so
gaps which has a size of nearly double('nearly is, depend on
contour form) of this value would be closed.