Sizechangemode: supports 'NON-PAINTING' devices.
Now sizechangemode supports which is configured as
'NON-PAINTING' behavior.

When NON-PAINTING device button pressed, button_press_cb
fired from gui.document. but, button_release_cb never
fired.
It is right behavior, because that device should not do
any of drawing.
But, sizechangemode does not draw anything...

so I wrote special workaround for it.