grabcutfill:Reconsider about multicolor painting.
Multiple colors hint turns out to be useless in most case.
It does not work like GMIC or LazyBrush.
However this tool is very experimental... but it is too bad.
I think that because OpenCV grabCut is basically for photographic
image, not for cartoon/anime like lineart.

So I reconsider and change grabcutfill as
'Only fill current brush color', i.e. only current brush colored
hints are recognized as Foreground hints.
and other colored hints are treated as Background hints.

There is still a lot of problem, but this might become useful tool??