grabcutfill: Added OpenCV2 based fill functionality
Grabcutfill tool, which use grabCut function of OpenCV2,
is added.

It needs python-opencv package, which is not included
official mypaint library dependency.
So that tool imports cv2 module when it is needed,
not ordinary import lines at the top of the file.

When grabcutfill tool is executed without python-opencv,
it just pops up a warning dialog and does nothing.