freehand_parallel: now consider 'rough' margin.
To make 'snap to level' button work in more rough level angle,
I added a concept of 'rough margin'.

When the ruler angle come to in the range of
'roughly margined' level, the ruler bar grow bright(washed)
green.
When the ruler approaches almost level, it changes to high
saturation green.

Even in a roughly level state, we can make it very accurate
level by pressing the 'snap to level' button.