Added 'CutLayerDown' functionality.
Added 'CutLayerDown' functionality, which does
'Cut current layer with the transparent area of benath'.
As a result, we can paint only opaque area of
the benath layer, without doing 'MergeLayerDownOpaque'
(i.e. continue painting as another layer)
This feature is useful when you want to paint reflection
into shadow of foreground objects.

Almost same as this can be done with combination mode of
'Source-atop', but it need to group those layers.
CutLayerDown can do it much easier, immidiately.