AutoModeChange: Disabled temporary.
`Auto Mode Change` by device is temporary disabled,
Because it might be confusing and rather harmful feature.

For example, you draw something with pen stlus and move cursor
outside of canvas, and then switch to mouse,
then choice `Move Layer` tool(mode) from icon,
and move pointer into canvas... the freehand tool would be
automatically selected unintentionally.
This is annoying, but we can not know which device is used
when icon button pressed.

So I might delete this feature.